requirements.txt
streamlit
pandas
langchain == 0.1.19
langchain-community == 0.0.38
langchain-core == 0.1.52
langchain-experimental == 0.0.57
langchain-openai == 0.1.6
faiss-cpu == 1.8.0
pydantic == 1.10.9
pandas == 2.2.2
emoji == 2.11.1
google-generativeai 
langchain-google-genai
requests 
python-dotenv 
flask==3.0.3
flask_cors==4.0.1
Flask-Session==0.8.0
pymupdf==1.24.9
scikit-learn==1.5.1
PyPDF2==3.0.1